pyproject.toml
[project]
name = "crew-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-huggingface>=1.2.1",
    "load-dotenv>=0.1.0",
]
